Adding a new job_config section to es-info dict.
1. Get the job config via Jenkins "get_job_config" api call.
2. Store the returned xml in "job_config_xml" attribute.
3. Look into the job config and attempt to interpret some Jenkins Pipeline job specific data.

If it is a pipeline job, return the "pipeline_job_type": "Script", "SCM" or "Multibranch".
If it is using SCM to define the job via a Jenkinsfile script, note the "jenkinsfile" name.
If it is using git as its SCM plugin, note the configured "git_repo" and the "git_branch".