Coloured Ansible output can break host status matching
When the ansible coloured output is parsed, an ansi escape sequence
is in the host line, space-separated from the status elements.

The split of this results in an element that does not have an '=' and
we fail to unpackj that split to the 2 waiting variables.

Resolve this by removing ANSI escape sequences from the console log.
Regex taklen from:
https://stackoverflow.com/questions/14693701/how-can-i-remove-the-ansi-escape-sequences-from-a-string-in-python