Avoid object/value error if config.xml is not retreived
If we don't have permissions to read the job configuration we set the
job_config_info to a string indicating so.  This actually prevents the
ingestion of the main build value as the job_config_info value might
previously have been read and interpreted as an object.

Instead, set the job_config_info to None and introduce the
job_config_info_status key to indicate status.